requirements.txt
streamlit
google-generativeai
python-dotenv
langchain-community
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
cassio
tiktoken
datasets
